requirements.txt
Django==1.9.5
Pillow==3.2.0
PyJWT==1.4.0
defusedxml==0.4.1
dj-database-url==0.4.1
django-address==0.1.5
django-allauth==0.25.2
django-conversation==1.4.6
django-libs==1.67.4
django-widget-tweaks==1.4.1
gunicorn==19.4.5
oauthlib==1.1.0
python-social-auth==0.2.19
python3-openid==3.0.10
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
whitenoise==3.0
psycopg2==2.5.4
